Clean up PR
